# packages/shared/openapi.yaml
openapi: 3.0.3
info:
  title: FM-Lab REST API
  description: REST API for programmatic access to FileMaker metadata stored in DuckDB
  version: 1.0.0
  contact:
    name: Marcel
  license:
    name: ISC
servers:
  - url: http://localhost:3003
    description: Local development server
...
paths:
...
  /reference/steps/{idOrSlug}/grammar:
    get:
      summary: XMLSnippet grammar of one script step (fm-spec emission layer)
      description: >-
        Snippet template, element order, options (with per-value evidence),
        structural constraints incl. the known-FM-bug registry kinds
        (clipboard_loss, version_skew, save_corruption,
        serialization_unstable, localized_build_defect — a warning class,
        never a validity rule), repeat groups (fm_spec >= 1.15.0;
        fixed-slot columns max_items/slot_positional/pad_mode >= 1.16.0),
        and the hint-inventory layer (fm_spec >= 1.17.0): skeleton hulls,
        option-bound elements, notation implications, the structural
        variable-target marker and the per-kind consumer note of the
        bug registry. Fields absent from older reference builds are
        null/empty — consumers must stay graceful.
      tags: [Reference]
      parameters:
        - name: idOrSlug
          in: path
          required: true
          schema:
            type: string
          description: Numeric step id or url_slug
      responses:
        '200':
          description: Step grammar (data envelope)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      stepId:
                        type: integer
                      canonicalName:
                        type: string
                      available:
                        type: boolean
                        description: false when the reference has no grammar row
                      xmlMap:
                        type: object
                        nullable: true
                        properties:
                          snippetTemplate: { type: string }
                          saxmlParamTypes: { type: string, nullable: true }
                          saxmlExample: { type: string, nullable: true }
                          elementOrder: { type: string, nullable: true }
                          variableTargetMarker:
                            type: boolean
                            nullable: true
                            description: >-
                              true = the snippet carries an empty <Text/>
                              step-level marker exactly when a target slot
                              holds a variable (fm_spec >= 1.17.0; null =
                              column absent on this reference build)
                          evidence: { type: string, nullable: true }
                          verifiedVersion: { type: string, nullable: true }
                          notes: { type: string, nullable: true }
                      options:
                        type: array
                        items:
                          type: object
                          properties:
                            optionKey: { type: string }
                            optionType: { type: string }
                            required: { type: boolean }
                            displayLocation: { type: string, nullable: true }
                            displayLabelEn: { type: string, nullable: true }
                            trueText: { type: string, nullable: true }
                            falseText: { type: string, nullable: true }
                            omitWhenFalse: { type: boolean }
                            invertedLabel: { type: boolean }
                            xmlPath: { type: string, nullable: true }
                            sortOrder: { type: integer, nullable: true }
                            evidence: { type: string, nullable: true }
                            verifiedVersion: { type: string, nullable: true }
                            values:
                              type: array
                              items:
                                type: object
                                properties:
                                  xmlValue: { type: string }
                                  displayTextEn: { type: string, nullable: true }
                                  evidence: { type: string, nullable: true }
                      constraints:
                        type: array
                        items:
                          type: object
                          properties:
                            constraintKind: { type: string }
                            detail: { type: string, nullable: true }
                            evidence: { type: string, nullable: true }
                            verifiedVersion: { type: string, nullable: true }
                            consumerNote:
                              type: string
                              nullable: true
                              description: >-
                                Consumer-facing lead text of the constraint
                                kind (fm_spec >= 1.17.0 constraint_kinds
                                registry; bug-registry kinds only)
                      repeatGroups:
                        type: array
                        description: Empty on references older than 1.15.0
                        items:
                          type: object
                          properties:
                            groupKey: { type: string }
                            groupLabel: { type: string }
                            parentGroup: { type: string, nullable: true }
                            containerPath: { type: string }
                            countAttr: { type: string, nullable: true }
                            itemForm:
                              type: string
                              enum: [bracket, scalar]
                            maxItems: { type: integer, nullable: true }
                            slotPositional: { type: boolean, nullable: true }
                            padMode: { type: string, nullable: true }
                            evidence: { type: string, nullable: true }
                            verifiedVersion: { type: string, nullable: true }
                      skeletonElements:
                        type: array
                        description: >-
                          Skeleton hulls the step keeps alive even
                          unconfigured (fm_spec >= 1.17.0; empty on older
                          references)
                        items:
                          type: object
                          properties:
                            parentTag: { type: string }
                            childTag: { type: string }
                            conditionOption: { type: string, nullable: true }
                            conditionValue: { type: string, nullable: true }
                            keepMode:
                              type: string
                              enum: [hull, hull_strip_children]
                            evidence: { type: string, nullable: true }
                            verifiedVersion: { type: string, nullable: true }
                      elementBindings:
                        type: array
                        description: >-
                          Option-value/element couplings (fm_spec >= 1.17.0;
                          empty on older references)
                        items:
                          type: object
                          properties:
                            optionKey: { type: string, nullable: true }
                            optionValue: { type: string, nullable: true }
                            elementPath: { type: string }
                            binding:
                              type: string
                              enum: [requires, excludes, requires_option,
                                     excludes_option, suppress_empty]
                            evidence: { type: string, nullable: true }
                            verifiedVersion: { type: string, nullable: true }
                      optionImplications:
                        type: array
                        description: >-
                          Parse-side notation implications (fm_spec >= 1.17.0;
                          empty on older references)
                        items:
                          type: object
                          properties:
                            triggerKind:
                              type: string
                              enum: [option_present, value_form, keyword,
                                     mode_switch]
                            trigger: { type: string }
                            impliedOption: { type: string }
                            impliedValue: { type: string, nullable: true }
                            isDefault: { type: boolean, nullable: true }
                            direction:
                              type: string
                              enum: [parse]
                            evidence: { type: string, nullable: true }
                            verifiedVersion: { type: string, nullable: true }
        '404':
          description: Unknown step id/slug (REF_STEP_NOT_FOUND)
...